Move sign up method to firebase class
